pyproject.toml
[tool.poetry]
name = "pykg"
version = "0.1.0"
description = ""
authors = ["Agung Setiaji <mragungsetiaji@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8"
google-cloud-dataproc = "2.0.2"
google-cloud-storage = "1.32.0"
loguru = "0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
